trade volume was based on daily sales instead of total

+++ src/main/java/lan/groland/eve/domain/market/Trade.java
@@ -6,7 +6,7 @@ public interface Trade extends Comparable<Trade> {
 
   double dailyBenefit();
 
-  int volume();
+  double volume();
 
   double capital();
